UPSTREAM: drm/dsi: Check for CONFIG_OF when defining of_mipi_dsi_device_add()
of_mipi_dsi_device_add() is used only when CONFIG_OF is enabled. It
currently works if OF support is disabled, but this will change when we
add more functionality to it.

Define the original function if CONFIG_OF is enabled and a dummy
function otherwise.

Signed-off-by: Archit Taneja <architt@codeaurora.org>
Signed-off-by: Thierry Reding <treding@nvidia.com>

Change-Id: I5815c8fe403b65989bfbaf33aba208d83634fa4a
Signed-off-by: Wyon Bi <bivvy.bi@rock-chips.com>
(cherry picked from commit fc903ebd50d0396b5de1ff614a009c2ddcb3222c)